refactor(editor): Move node issue functions from useNodeHelpers to useWorkflowState
This removes the dependency of useNodeHelpers on useWorkflowState by:

1. Creating nodeIssueUtils.ts with shared utilities (hasProxyAuth, displayParameter)
2. Creating useWorkflowNodeIssuesState composable with all node issue functions
3. Spreading node issues functions into useWorkflowState return object
4. Updating all consumer files to use workflowState instead of nodeHelpers

Functions moved to workflowState:
- updateNodeInputIssues, updateNodesInputIssues, updateNodesExecutionIssues
- updateNodeParameterIssues, updateNodesParameterIssues
- updateNodeCredentialIssues, updateNodeCredentialIssuesByName
- updateNodeParameterIssuesByName, updateNodesCredentialsIssues
- disableNodes, getNodeIssues, getForeignCredentialsIfSharingEnabled
